+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
-name = "ddn"
-version = "3.0.0"
+name = "ddn3"
+version = "0.1.0"
 description = "Dependency network analysis under two conditions"
 authors = ["Minta Lu", "Yizhi Wang"]
 license = "MIT"
