+++ src/ddn3/strong_rule.py
